+++ src/module/sprintf_s.c
@@ -4,7 +4,7 @@
 #ifdef __GNUC__
 /** Case-insensitive memcmp() */
 int 
-memcasecmp(const void *lhs, const void *rhs, int length)
+memcasecmp(const void *lhs, const void *rhs, size_t length)
 {
     int i;
     for (i=0; i<length; i++) {
